% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BlandAltman}
\alias{BlandAltman}
\title{Produce a Bland-Altman style plot for agreement between numeric measures.}
\usage{
BlandAltman(data, x1, x2, alpha = 0.05, gg_expr = list())
}
\arguments{
  \item{data}{A data frame.}

  \item{alpha}{The significance level, or 1 - the
  confidence level.}

  \item{gg_expr}{A list of additional layers to be added to
  the \code{ggplot} object.  For instance,
  \code{facet_grid()} can be an element of the list, or
  \code{ggtitle}, etc.}

  \item{x}{a numeric vector giving one set of values for
  comparison}

  \item{y}{a numeric vector giving the second set of values
  for comparison}
}
\description{
The Bland-Altman plot places the difference between two
measures (such as observations between two readers) on the
y-axis against the pair-wise means of the observations on
the x-axis.  The confidence limits of the differences are
plotted.  A plot of good agreement should have a random
dispersal of points within the confidence limits, and at
least 95% of the points should be within those limits.
Using \code{ggplot}, these plots can be generated in a grid
with each panel having limits calculated independently of
the others.
}
\details{
The use of \code{colour}, \code{shape}, and other
\code{ggplot} aesthetics is not recommended with this
function.  Any data within the panel will be included in
the calculation of the confidence limits.  If additional
aesthetics are being used to show strata of data, these
confidence limits will not be correctly calculated.
}
\author{
Benjamin Nutter
}

