% Generated by roxygen2 (4.0.2): do not edit by hand
\name{copyTemplate}
\alias{copyTemplate}
\title{Copy Analysis File Template to a Directory}
\usage{
copyTemplate(templateName, renameAs, copyToDir = getwd(),
  listTemplates = FALSE, ext = c("rmd", "rnw", "rhtml", "tex"), ...)
}
\arguments{
  \item{renameAs}{A character string giving the new name of
  the file in the new location.}

  \item{copyToDir}{Location to which the template should be
  copied.}

  \item{ext}{The file extension for the template.}

  \item{...}{Additional arguments to pass to
  \code{file.copy}.}

  \item{templatename}{The file name of the template.  The
  file extension should be left off.  Note that there may
  be up to four templates for each template name, depending
  on the extension chosen in \code{ext}.}

  \item{listTemplate}{Logical.  If \code{TRUE}, a list of
  all available templates will be written to the console.
  This does not interrupt the normal process of the
  function.}
}
\description{
Various templates are provided as a part of
\code{CCFproprietary} that fit the needs of various
individuals or departments.  \code{copyTemplate} allows the
user to copy these templates into a project directory
without having to manually 'Save As' a file.
}
\examples{
\dontrun{
copyTemplate("nutter", "01-StudyDesign",
             copyToDir="s_plus")
}
}
\author{
Benjamin Nutter
}

