% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findVariable}
\alias{findVariable}
\title{Find data frames containing a selected variable names}
\usage{
findVariable(var)
}
\arguments{
  \item{var}{A vector of variable names to search for.
  Only data frames containing \emph{all} of the variables
  given will be returned.}
}
\description{
The data frames in a workspace are searched for variable
names of interest and a vector of names is returned.  The
returned vector names the data frames that contain all of
the variables of interest.
}
\details{
This function was written in response to the following
query:

"I have a workspace attached in R.  it's got objects in it.
some of the objects are data frames.  I'd like to search
the dataframes in the workspace for a variable called THGRP
and get a listing of dataframes that contain it.  how I do
this?"
}
\examples{
mtcarsA <- mtcars
mtcarsB <- mtcars
mtcarsB$newVar <- 1:nrow(mtcarsB)
IndomethA <- Indometh

findVariable("am")
findVariable(c("am", "mpg"))
findVariable(c("am", "newVar"))
findVariable("time")

# cleanup
rm(list=c('mtcarsA', 'mtcarsB', 'IndomethA'))
}
\author{
Benjamin Nutter
}

