% Generated by roxygen2 (4.0.2): do not edit by hand
\name{interval_2p}
\alias{interval_2p}
\title{Sample Size for the confidence interval of the differnce of proportions.}
\usage{
interval_2p(E = NULL, p1 = NULL, p2 = NULL, n = NULL, n1 = NULL,
  n2 = NULL, k = 0.5, alpha = 0.05)
}
\arguments{
  \item{E}{Margin of error for the difference of
  proportions}

  \item{p1}{Proportion observed in group 1.}

  \item{p2}{Proportion observed in group 2.}

  \item{n}{Total sample size for groups 1 and 2.}

  \item{n1}{Sample size for group 1.}

  \item{n2}{Sample size for group 2.}

  \item{k}{proportion of the total samples size \code{n}
  that consitutes \code{n1}. the default assumptions is
  that \code{n1 = n2}.}

  \item{alpha}{Significance level (or 1-confidence)}
}
\description{
The $z$-based confidence interval is commonly used in
estimating population mean using sample data.  It is based
on the Standard Normal distribution.  \code{interval_2p}
provides tools for calculating and illustrating the sample
size requirement and other design factors associated with
these intervals.
}
\details{
This function is currently under development.  I needed it
for a study design.

This can be useful for designin equivalence trials.
According to Schumi and Wittes, ``Sometimes, the goal is
not to show that the new treatment is better, but that the
new treatment is 'equivalent' to the control...At the end
of the trial, a CI is computed around the difference
between two test statistics (equivalence trials typically
use 90% CIs) and if the CI lies strictly within [-E, +E]
the two treatments are called 'equivalent.'"
}
\author{
Benjamin Nutter
}
\references{
Jennifer Schumi and Janet T Wittes, `` Through the looking
glass: understanding non-inferiority," _Trials_, 2011,
12:106.
}

