% Generated by roxygen2 (4.0.2): do not edit by hand
\name{interval_auc_1curve}
\alias{interval_auc_1curve}
\title{Single Predictor, Single Curve ROC Confidence Interval Sample Sizes}
\usage{
interval_auc_1curve(auc = NULL, n = NULL, E = NULL, alpha = 0.05,
  weights = list(c(1, 1)), optim.max = 1e+09)
}
\arguments{
  \item{auc}{Observed Area Under the Curve (AUC)}

  \item{n}{Total sample size}

  \item{E}{Desired margin of error}

  \item{alpha}{Significance Level}

  \item{weights}{a list of length-2 vectors giving the
  weights for the diseased and non-diseased groups.  This
  is used to calculate \code{k}, the proportion of subjects
  belonging to the first group.}

  \item{optim.max}{Maximum value for \code{optimze} when
  solving for \code{n}.  The default setting is
  1,000,000,000, which is probably higher than anyone will
  ever need.  I may change this in the future for the sake
  of improving speed.}
}
\value{
Returns a data frame of the parameters passed to the
function (expanded using \code{expand.grid}) and the
corresponding sample size estimates.

\enumerate{ \item \code{E}: Margin of Error \item
\code{auc}: Area Under the Curve (AUC) \item \code{k}:
Proportion of total sample size belonging to group 1 \item
\code{n1_est}: Exact sample size for group 1 \item
\code{n2_est}: Exact sample size for group 2 \item
\code{n_est}: Exact total sample size \item \code{n1}:
Estimated sample size for group 1 (next largest integer)
\item \code{n2}: Estimated sample size for group 2 (next
largest integer) \item \code{n}: Estimated total sample
size (sum of \code{n1} and \code{n2}) }
}
\description{
Calculates the sample size, AUC, margin of error, or
significance for the confidence interval of an Area Under
the Curve (AUC). Assumes that the AUC is based on a single
numeric predictor and that the AUC is normally distributed.
}
\details{
Exactly one of the parameters \code{n}, \code{E},
\code{auc} and \code{alpha} must be passed as \code{NULL},
and that parameter will be calculated from the others.
Notice that \code{alpha} has a non-\code{NULL} default, so
\code{NULL} must be explicitly passed if you want it
computed.
}
\examples{
# (Example 1 from PASS Sample Size Software)
# \\url{http://ncss.wpengine.netdna-cdn.com/wp-content/themes/ncss/pdf/Procedures/PASS/Confidence_Intervals_for_the_Area_Under_an_ROC_Curve.pdf}
interval_auc_1curve(auc=c(0.6, 0.7, 0.8, 0.9),
                    E = c(.05, .10),
                    alpha=0.05)
}
\author{
Benjamin Nutter
}
\references{
Nancy A Obuchowski, "Sample size calculations in studies of
test accuracy," \emph{Statistical Methods in Medical
Research} 1998; 7: 371-392

James A. Hanley and Barbara J. McNeil, "The Meaning and Use
of the Area under a Receiver Operating Characteristic (ROC)
Curve," \emph{Radiology}. Vol 143. No 1, Pages 29-36, April
1982.
}

