% Generated by roxygen2 (4.0.2): do not edit by hand
\name{interval_auc_2curve}
\alias{interval_auc_2curve}
\title{Single Predictor, Difference Between Two ROC Curve Confidence Interval Sample Sizes}
\usage{
interval_auc_2curve(auc1 = NULL, auc2 = NULL, n = NULL, E = NULL,
  alpha = 0.05, r, weights = list(c(1, 1)), optim.max = 1e+09)
}
\arguments{
  \item{auc1}{Observed Area Under the Curve 1 (AUC)}

  \item{auc2}{Observed Area Under the Curve 2 AUC)}

  \item{n}{Total sample size}

  \item{E}{Desired margin of error}

  \item{alpha}{Significance Level}

  \item{r}{The correlation introduced between the two AUCs
  by studying the same sample of patients.}

  \item{weights}{a list of length-2 vectors giving the
  weights for the diseased and non-diseased groups.  This
  is used to calculate \code{k}, the proportion of subjects
  belonging to the first group.}

  \item{optim.max}{Maximum value for \code{optimze} when
  solving for \code{n}.  The default setting is
  1,000,000,000, which is probably higher than anyone will
  ever need.  I may change this in the future for the sake
  of improving speed.}
}
\value{
Returns a data frame of the parameters passed to the
function (expanded using \code{expand.grid}) and the
corresponding sample size estimates.

\enumerate{ \item \code{E}: Margin of Error \item
\code{auc1}: Area Under the Curve 1 (AUC) \item
\code{auc2}: Area Under the Curve 2 (AUC) \item \code{k}:
Proportion of total sample size belonging to group 1 \item
\code{r}: The correlation introduced between the two AUCs
by studying the same sample of patients.  \item
\code{n1_est}: Exact sample size for group 1 \item
\code{n2_est}: Exact sample size for group 2 \item
\code{n_est}: Exact total sample size \item \code{n1}:
Estimated sample size for group 1 (next largest integer)
\item \code{n2}: Estimated sample size for group 2 (next
largest integer) \item \code{n}: Estimated total sample
size (sum of \code{n1} and \code{n2}) }
}
\description{
Calculates the sample size, AUC's, margin of error, or
significance for the confidence interval of the difference
of two Areas Under the Curve (AUC). Assumes that the AUCs
are based on a single numeric predictor and that the AUCs
are normally distributed.
}
\details{
The ROC curves for this comparison should be based on the
same sample.  That is, it should be two tests performed on
the same sample.  The comparison assumes some dependence
between the curves, with this dependence being defined by
\code{r}.

Exactly one of the parameters \code{n}, \code{E},
\code{auc1}, \code{auc2} and \code{alpha} must be passed as
\code{NULL}, and that parameter will be calculated from the
others.  Notice that \code{alpha} has a non-\code{NULL}
default, so \code{NULL} must be explicitly passed if you
want it computed.
}
\author{
Benjamin Nutter
}
\references{
James A Hanley and Barbara J McNeil, "A Method of Comparing
the Areas Under Receiver Operating Characteristic Curves
Derived from the Same Cases," \emph{Radiology}, Vol. 148,
No. 3, Pages 839-843, September 1983.

Nancy A Obuchowski, "Sample size calculations in studies of
test accuracy," \emph{Statistical Methods in Medical
Research} 1998; 7: 371-392

James A. Hanley and Barbara J. McNeil, "The Meaning and Use
of the Area under a Receiver Operating Characteristic (ROC)
Curve," \emph{Radiology}. Vol 143. No 1, Pages 29-36, April
1982.
}

