% Generated by roxygen2 (4.0.2): do not edit by hand
\name{n_t1samp_interval}
\alias{n_t1samp_interval}
\title{Sample Size for One Sample T Confidence Intervals}
\usage{
n_t1samp_interval(E = NULL, n = NULL, s = NULL, alpha = 0.05,
  optim.max = 1e+09)
}
\arguments{
  \item{E}{Margin of error}

  \item{n}{Sample size}

  \item{s}{Sample standard deviation}

  \item{alpha}{Significance level (or 1-confidence)}

  \item{optim.max}{Maximum value for \code{optimze} when
  solving for \code{n}.  The default setting is
  1,000,000,000, which is probably higher than anyone will
  ever need.  I may change this in the future for the sake
  of improving speed.}
}
\value{
Returns an object of class \code{HazPwr} with subclasses
\code{t1samp} and \code{est} (actual appearance:
\code{t1samp_est_HazPwr}.  The object contains a data frame
of the parameters passed to the function (expanded using
\code{expand.grid}) and the corresponding sample size
estimates.

\enumerate{ \item \code{n_est}: Exact sample size.  When
\code{n} is not \code{NULL}, this will be identical to
\code{n}.  \item \code{n}: Estimated sample size; the next
largest integer after \code{n_est}.  \item \code{E}: Margin
of error \item \code{s}: Sample Standard Deviation \item
\code{alpha}: Desired significance level.  }
}
\description{
The $t$-based confidence interval is commonly used in
estimating population mean using sample data.  It is based
on Student's $t$ distribution, an adaptation of the
standard normal distribution.  \code{n_t1samp_interval}
provides tools for calculating and illustrating the sample
size requirement and other design factors associated with
these intervals.
}
\details{
Exactly one of the parameters \code{n}, \code{E}, \code{s}
and \code{alpha} must be passed as \code{NULL}, and that
parameter will be calculated from the others.  Notice that
\code{alpha} has a non-\code{NULL} default, so \code{NULL}
must be explicitly passed if you want it computed.
}
\author{
Benjamin Nutter
}

