% Generated by roxygen2 (4.0.2): do not edit by hand
\name{odds_noninferior}
\alias{odds_noninferior}
\title{Sample Size for a non-inferiority test of an Odds Ratio.}
\usage{
odds_noninferior(pA = NULL, pB = NULL, delta = NULL, kappa = 1,
  alpha = 0.05, power = NULL)
}
\arguments{
  \item{pA}{}

  \item{pB}{}
}
\description{
Make a good description here.
}
\details{
This function is currently under development.  I needed it
for a study design.
}
\author{
Benjamin Nutter
}
\references{
http://powerandsamplesize.com/Calculators/Test-Odds-Ratio/Non-Inferiority-or-Superiority
}

