% Generated by roxygen2 (4.0.2): do not edit by hand
\name{password}
\alias{password}
\title{Random Password Generator}
\usage{
password(length = 8, max.to.num = length - 1, use.num = TRUE)
}
\arguments{
  \item{length}{Character length of the password.}

  \item{max.to.num}{maximum number of characters that may
  be selected before choosing a numeric character.}

  \item{use.num}{Logical, determines if numeric characters
  may be used.}
}
\description{
Generates a password consisting of a random set of
characters.  Passwords are generated in a way that prevents
any more than two consecutive characters to be typed by the
same hand.
}
\details{
At present, special (punctuation) characters and capital
letters are not among the sampled characters.
}
\examples{
password()
}
\author{
Benjamin Nutter
}

