% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Power.t.test}
\alias{Power.t.test}
\title{Modified T-test Power Calculations}
\usage{
Power.t.test(n = NULL, delta = NULL, sd = 1, sig.level = 0.05,
  power = NULL, type = "two.sample", alternative = "two.sided",
  strict = FALSE)
}
\arguments{
  \item{n}{Number of observations (per group)}

  \item{delta}{True difference in means}

  \item{sd}{Standard deviation}

  \item{sig.level}{Significance level (Type I error
  probability)}

  \item{power}{Power of test (1 minus Type II error
  probability)}

  \item{type}{Type of test}

  \item{alternative}{One or two-sided test}

  \item{strict}{Use strict interpretation in two sided
  case}
}
\description{
Allows for calculating power, effect size, or sample size
over multiple conditions.  Returns a data frame instead of
an 'htest' object
}
\details{
This is essentially a wrapper for a for loop over
\code{power.t.test}.
}

