% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchFunctions}
\alias{searchFunctions}
\title{Search Functions for Character Strings}
\usage{
searchFunctions(pattern, directory)
}
\arguments{
  \item{pattern}{The character string to be matched}

  \item{directory}{The directory of functions to be
  scanned}
}
\description{
Search Functions for Character Strings
}
\details{
At times, I have needed to find certain character strings
within a package.  For instance, I may need to know which
functions contain a particular argument.  This function
scans all of the files in a directory for a character
string and returns the names of the files that contain it.
}
\examples{
\dontrun{
searchFunctions("survival",
                "C:/Users/nutterb/Documents/GitHub/lazyWeave/R")
}
}
\author{
Benjamin Nutter
}

