% Generated by roxygen2 (4.0.2): do not edit by hand
\name{survTable}
\alias{survTable}
\title{Univariable Survival Summary of Categorical Variables}
\usage{
survTable(data, vars, byVar, timeVar, times = 60)
}
\arguments{
  \item{data}{A data frame with the variable to be
  summarized}

  \item{vars}{A character vector of categorical variables
  to be used as predictors of survival}

  \item{byVar}{A character(1) vector of 0's and 1's
  indicating whether the subject is censored or experienced
  the event}

  \item{timeVar}{A character(1) vector giving the name of
  the variable with the survival time}

  \item{times}{A numeric(1) vector giving the time at which
  the mean survival should be calculated.}
}
\description{
Provides the median (and CI) survival, mean (and CI)
survival, and hazard ratios (and CI) for time to event
analyses related to categorical variables.
}
\details{
This function is under active development and may change
without notice.
}
\author{
Benjamin Nutter
}

