% Generated by roxygen2 (4.0.2): do not edit by hand
\name{survTableComplex}
\alias{survTableComplex}
\title{Univariable Summary of Survival Analyses}
\usage{
survTableComplex(data, vars, byVar, time.var, times = NULL, vars.cat,
  event.level, normal = NULL, quartile = TRUE, row.p = FALSE)
}
\arguments{
  \item{data}{A data frame}

  \item{vars}{Variables to be summarized}

  \item{byVar}{The event variable.  May be 0/1 or a two
  category factor/character}

  \item{time.var}{Variable giving the survival time}

  \item{times}{The times argument passed to
  \code{summary.survfit}}

  \item{vars.cat}{Numeric variables in 'vars' to force as
  categorical variables}

  \item{event.level}{For non-numeric \code{byVar},
  specifies the level to be considered the event.}

  \item{normal}{a character vector naming numeric variables
  to be summarized by mean and standard deviation.}

  \item{quartile}{Should non-normal numeric variables be
  summarized with the interquartile range. If \code{FALSE},
  the range is printed.}

  \item{row.p}{For categorical variables, should the
  percentage show the row percentage?}
}
\description{
Provides a summary table of univariable survival metrics.
The default summaries include summary values similar to
\code{cattable} and \code{conttable}, but also includes the
hazard ratio, p-value, and the median survival.
Additionally, the survival proportions for user specified
time points may be given.
}
\details{
The table produced is similar to the \code{cattable}
format, but the output may not be as flexible.
}
\author{
Benjamin Nutter
}

