% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_auc_1curve}
\alias{test_auc_1curve}
\title{Power Analsis for Full Area Under the ROC Curve}
\usage{
test_auc_1curve(auc0 = NULL, auc1 = NULL, n = NULL, power = NULL,
  alpha = 0.05, weights = list(c(1, 1)), one_tail = FALSE,
  ordinal = FALSE)
}
\arguments{
  \item{auc0}{The hypothesize (null) value of AUC.}

  \item{auc1}{The observed or desired AUC.}

  \item{n}{The total sample size}

  \item{power}{The power of the test}

  \item{alpha}{The significance level of the test.}

  \item{weights}{Weighting of subjects with the condition
  and those without. A vector \code{1, 3} indicates there
  is one subject with the condition for every three
  subjects without it.  Defaults to equal sizes.  The
  weights are normalized, so they do not have to sum to
  one.}

  \item{one_tail}{Logical. Indicates if the test should be
  a one tailed test.}

  \item{ordinal}{Logical.  Indicates if the independent
  variable in the study is ordinal.  This is currently
  ignored.}
}
\description{
Calculates the power, sample size, or other parameters for
studies using Area Under the Curve (AUC) as the endpoint of
an anlysis using a single continuous or ordinal predictor.
}
\details{
The Hanley and McNeil estimator is used for continuous
predictors.  Obuchowski's estimator for discrete ratings is
used for ordinal predictors.
}
\examples{
#* First example from \\url{http://www.bio.ri.ccf.org/doc/rocpower_help.txt}
#* but solving for n
test_auc_1curve(auc0 = .75, auc1 = .92,
                n = NULL, alpha=.02,
                power=.893, weights = list(c(35, 50)))
#* First example from \\url{http://www.bio.ri.ccf.org/doc/rocpower_help.txt}
test_auc_1curve(auc0 = .75, auc1 = .92,
                n = 85, alpha=.02,
                power=NULL,
                weights = list(c(35, 50)))
}
\author{
Benjamin Nutter
}
\references{
Nancy A Obuchowski, "Sample size calculations in studies of
test accuracy," \emph{Statistical Methods in Medical
Research} 1998; 7: 371-392

James A. Hanley and Barbara J. McNeil, "The Meaning and Use
of the Area under a Receiver Operating Characteristic (ROC)
Curve," \emph{Radiology}. Vol 143. No 1, Pages 29-36, April
1982.

The bulk of this code is compared against C Zepp's SAS
Macro ROCPOWER. See reference 25 in the Obuchowski paper
above.
\url{http://www.bio.ri.ccf.org/doc/rocpower_help.txt}
\url{http://www.bio.ri.ccf.org/doc/rocpower.sas}
}

