% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_logrank_hr_1binary}
\alias{test_logrank_hr_1binary}
\title{Power Analysis for Log Rank Tests with a Single Binary Covariate}
\usage{
test_logrank_hr_1binary(hr = NULL, deaths = NULL, p = 0.5, alpha = 0.05,
  power = 0.8)
}
\arguments{
  \item{hr}{Hazard ratio}

  \item{deaths}{Number of deaths}

  \item{p}{proportion of subjects to be randomized to the
  first group.  The format of this argument will eventually
  change.}

  \item{alpha}{significance level for the test}

  \item{power}{Power of the test}
}
\description{
Calculates the power, number of deaths, proportion to be
randomized, significance, or hazard ratio for a log-rank
test of a Kaplan-Meier curve with a single binary
covariate.
}
\details{
This function is currently in development and will undergo
serious changes. Among the changes: 1) p will be changed to
a 'weights' style argument; 2) follow up time and
recruitment interval will be incorporated; 3) event rates
may be incorporated so as to calculate the sample size
instead of the number of deaths.
}
\examples{
#* Example from Schoenfeld (doesn't incorporate recruitment and follow-up periods)
test_logrank_hr_1binary(hr=1.5, p=.25)
}
\author{
Benjamin Nutter
}
\references{
Schoenfeld DA. Sample-size formula for the
proportional-hazards regression model. \emph{Biometrics}
1983;39:499-503
}

